Fixed multi paged merging issue
